Paxos

This package implements the Paxos algorithm.

Contents

Paxos

Contents

Modules

Index

Modules

Paxos.Leaders — Module

Implementation of the leader side of the Paxos algorithm.

Paxos.Ledgers — Module

Data structure useful for tracking the state of one or more instances of the Paxos algorithm.

Paxos.Nodes — Module

Basic types for identifying nodes participating in Paxos

Paxos.Protocols — Module

Basic messages in the Paxos protocol.

Paxos.Transports.Common — Module

Defintions common to all implementations of Transport.

Paxos.Transports.Memory — Module

Implementation of Paxos.Transports.Common.Transport for in-memory communication within the current process.

Paxos.Transports.TCP — Module

Implementation of Paxos.Transports.Common.Transport for TCP-based communication between participants.

Paxos.Utils — Module

Utility functions for Paxos

Index

Paxos.Ballots

Paxos.Clients

Paxos.Configurations

Paxos.Leaders

Paxos.Ledgers

Paxos.Nodes

Paxos.Protocols

Paxos.Transports.Common

Paxos.Transports.Memory

Paxos.Transports.TCP

Paxos.Utils

Paxos.Configurations.Configuration

Paxos.Ballots.InstanceID

Paxos.Ballots.Promise

Paxos.Configurations.NoQuorumAvailableException

Paxos.Configurations.TransitionConfiguration

Paxos.Ledgers.Ledger

Paxos.Ledgers.LedgerEntry

Paxos.Ledgers.LedgerEntryState

Paxos.Nodes.NodeID

Paxos.Transports.Common.Connection

Paxos.Transports.Common.Listener

Paxos.Transports.Common.ListenerClosedException

Paxos.Transports.Common.Message

Paxos.Transports.Common.Messenger

Paxos.Transports.Common.Messenger

Paxos.Transports.Common.NoAvailableServerException

Paxos.Transports.Common.Transport

Paxos.Utils.TimeoutException

Base.close

Base.isready

Paxos.Ballots.after

Paxos.Ballots.before

Paxos.Clients.invoke

Paxos.Configurations.configuration

Paxos.Configurations.isquorum

Paxos.Configurations.isquorum

Paxos.Configurations.memberAddress

Paxos.Configurations.memberAddresses

Paxos.Configurations.quorumSize

Paxos.Configurations.quorumSize

Paxos.Leaders.prepareBallots

Paxos.Leaders.proposeBallots

Paxos.Ledgers.nextBallotNumber!

Paxos.Ledgers.nextInstance

Paxos.Ledgers.prepare!

Paxos.Ledgers.votes!

Paxos.Protocols.pcall

Paxos.Transports.Common.call

Paxos.Transports.Common.callAny

Paxos.Transports.Common.connectAll

Paxos.Transports.Common.connectOne

Paxos.Transports.Common.connectTo

Paxos.Transports.Common.connectTo

Paxos.Transports.Common.connection

Paxos.Transports.Common.gcall

Paxos.Transports.Common.listenOn

Paxos.Transports.Common.listenOn

Paxos.Transports.Common.listener

Paxos.Transports.Common.messageReceiver

Paxos.Transports.Common.messageSender

Paxos.Transports.Common.messengerTo

Paxos.Transports.Common.receiveFrom

Paxos.Transports.Common.receiveFrom

Paxos.Transports.Common.receivedMessages

Paxos.Transports.Common.sendMessage

Paxos.Transports.Common.sendTo

Paxos.Transports.Common.sendTo

Paxos.Transports.Memory.memory

Paxos.Utils.bounded

Paxos.Utils.compact

Paxos.Utils.groupBy

Paxos.Utils.identity

Paxos.Utils.maxBy

Paxos.Utils.readAvailable