Paxos.Configurations — Module

Data types and functions support Configurations, which describe the participants in a Paxos cluster.

Contents

Contents

Functions

Macros

Types

Constants

Index

Functions

Paxos.Configurations.configuration — Function

Create a new, empty OrdinaryConfiguration

Paxos.Configurations.isquorum — Method

Return true if the list of members represents a quorum for the configuration

Paxos.Configurations.isquorum — Method

During transitions, eturn true if the list of members represents a quorum in both old and new configurations

Paxos.Configurations.memberAddress — Method

Return the address for the specfied node, or error if no such address

Paxos.Configurations.memberAddresses — Method

Return the addresses of all membes in the configuration

Paxos.Configurations.quorumSize — Method

For Integer values, return the required number of participants for a majority

Paxos.Configurations.quorumSize — Method

Return the minimum size for a majority of members in the configuration

Macros

Types

Paxos.Configurations.NoQuorumAvailableException — Type

Exception thrown when no quorum is available to respond to Paxos messages

Paxos.Configurations.TransitionConfiguration — Type

A transition configuration (inspired by Raft) supports rollout of a new configuration, with potentially differences in membership

Constants

Paxos.Configurations.Configuration — Constant

A configuration is the general type for maintaining a list of menbers and addreses where they can be reached over transport

Index

Paxos.Configurations.Configuration

Paxos.Configurations.NoQuorumAvailableException

Paxos.Configurations.TransitionConfiguration

Paxos.Configurations.configuration

Paxos.Configurations.isquorum

Paxos.Configurations.isquorum

Paxos.Configurations.memberAddress

Paxos.Configurations.memberAddresses

Paxos.Configurations.quorumSize

Paxos.Configurations.quorumSize