Contents

Contents

Functions

Macros

Types

Constants

Index

Functions

Paxos.Ledgers.nextBallotNumber! — Function

Create a ballot number for an instance

Paxos.Ledgers.nextInstance — Method

Return the next unused instance (actually, the next unused index) in the ledger

Paxos.Ledgers.prepare! — Method

Record choices for each instance in ballots

Paxos.Ledgers.votes! — Method

Compute a ballot number corresponding to the highest sequence number seen for each instance. This is useful for the "propose" phase of the protocol.

Base.isready — Method

Return true if the ledger is not empty and has unapplied entries, false otherwise

Macros

Types

Paxos.Ledgers.Ledger — Type

A ledger is a record of Commands to apply to an external data structure or state machine. A ledger is structured as a sequence of LedgerEntry objects.

Paxos.Ledgers.LedgerEntry — Type

A LedgerEntry captures the state of a Command to apply to a Ledgers external model or state machine. Conveniently, a ledger entry has a structure suitable for use as a "ledger" for 1 instance of the Paxos algorithm: as ballots for that instance progress the entry contains the state needed for a leader, follower, or learner to move towards an outcome or final ballot containing the chosen command for application.

Paxos.Ledgers.LedgerEntryState — Type

open - The entry is new, and has not been populated by any ballot activity

promised- Once promised, only requests with a higher ballot can replace the request in the entry

accepted - An accepted entry will eventually by accepted by all members of the cluster

applied - The underlying command in the entry's request has been applied to application state

Constants

Index

Paxos.Ledgers.Ledger

Paxos.Ledgers.LedgerEntry

Paxos.Ledgers.LedgerEntryState

Paxos.Ledgers.nextBallotNumber!

Paxos.Ledgers.nextInstance

Paxos.Ledgers.prepare!

Paxos.Ledgers.votes!