Contents

Contents

Functions

Macros

Types

Constants

Index

Functions

Paxos.Transports.Common.call — Method

Send a message through the specified Messenger, and wait up to timeout seconds for a response. If no response received in the alloted time, then throw a TimeoutException; otherwise, return the reeived response.

Paxos.Transports.Common.callAny — Method

Call the messengers in order with the message, returning after the first messenger that responds without an exception. No other messengers after the first successful call will receive the message.

Paxos.Transports.Common.connectAll — Method

Asynchronously connect to as many destination addresses as possible, ignoring exceptions if encountered

Paxos.Transports.Common.connectOne — Method

Create a connection to the address and if successful retain it in connections

Paxos.Transports.Common.connectTo — Method

Create a connection (if one is possible) that enables sending and receiving messages to the recipient. The provided handler should take 1 argument, a Connection for sending or receiving messages with the recipient.

Paxos.Transports.Common.connection — Method

Create a Connection to the recepient and a Messenger to exchange messages, then invoke the handler with the Messenger as an argument. Close the messenger as soon as the connection exits.

Paxos.Transports.Common.gcall — Method

Send the message through each of the Messengers, but only allow timeout seconds for a response. Return all results received within the alloted time, if any.

Paxos.Transports.Common.listenOn — Method

Listen for incoming connections and invoke the indicated function when they appear. Returns a Listener. The provided handler should take 1 argument, a Connection; a new handler will be created for each connection accepted by the listener. A listener

Paxos.Transports.Common.listener — Method

Create a listener on the underlying transport at the indicated address, invoking the handler for each Connection discveroed. A listner should be closed when no longer needed, to free up resources

Paxos.Transports.Common.messengerTo — Method

Connect to the recipient via the underlying transport and construct a Messenger with the resulting Connection

Paxos.Transports.Common.receivedMessages — Method

Return a channel for processing messages received on the Messenger

Paxos.Transports.Common.sendMessage — Method

Enqueue message for sending over the transport connection

Paxos.Transports.Common.sendTo — Method

Send a message on the indicated connection

Base.close — Method

Close the connection, freeing resources associated with it–and stopping any tasks moving data on the connection.

Paxos.Transports.Common.messageReceiver — Method

Continuously receive messages over the connection and places them in the messenger's inbound channel, until the messenger is closed.

Paxos.Transports.Common.messageSender — Method

Continously send any messages on the messenger's outbound channel, until the messenger is closed

Paxos.Transports.Common.receiveFrom — Method

Receive a message over the indicated connection

Macros

Types

Paxos.Transports.Common.Connection — Type

A Connection represents a conduit between a source and a destination: messages may be sent to the destination from the source, and optionally they may also be received from the destination to the source. Closing a connection prevents further message exchange using the connection.

Paxos.Transports.Common.Listener — Type

A listener is typically a task that is awaiting incoming connections; closing a listener will prevent further connections to that listener.

Paxos.Transports.Common.Message — Type

An abstract payload for sending data via a Transport

Paxos.Transports.Common.Messenger — Type

A messenger handles asynchronous transmission / reception of messages over a Connection, until closed. In addition to stopping the tasks for transmission / reception, closing the messenger also closes the underlying connection.

Paxos.Transports.Common.Messenger — Method

Construct a messenger for the given resource

Paxos.Transports.Common.Transport — Type

A Transport delivers messages sent to groups or individual recipients

Paxos.Transports.Common.ListenerClosedException — Type

Exception sent to a listener to indicate that it has been closed

Paxos.Transports.Common.NoAvailableServerException — Type

Exception thrown when there is no server that can respond to a client request

Constants

Index

Paxos.Transports.Common.Connection

Paxos.Transports.Common.Listener

Paxos.Transports.Common.ListenerClosedException

Paxos.Transports.Common.Message

Paxos.Transports.Common.Messenger

Paxos.Transports.Common.Messenger

Paxos.Transports.Common.NoAvailableServerException

Paxos.Transports.Common.Transport

Paxos.Transports.Common.call

Paxos.Transports.Common.callAny

Paxos.Transports.Common.connectAll

Paxos.Transports.Common.connectOne

Paxos.Transports.Common.connectTo

Paxos.Transports.Common.connectTo

Paxos.Transports.Common.connection

Paxos.Transports.Common.gcall

Paxos.Transports.Common.listenOn

Paxos.Transports.Common.listenOn

Paxos.Transports.Common.listener

Paxos.Transports.Common.messageReceiver

Paxos.Transports.Common.messageSender

Paxos.Transports.Common.messengerTo

Paxos.Transports.Common.receiveFrom

Paxos.Transports.Common.receiveFrom

Paxos.Transports.Common.receivedMessages

Paxos.Transports.Common.sendMessage

Paxos.Transports.Common.sendTo

Paxos.Transports.Common.sendTo