Contents

Contents

Functions

Macros

Types

Constants

Index

Functions

Paxos.Transports.Common.connectTo — Method

Create a connection (if one is possible) that enables sending and receiving messages to the recipient. The provided handler should take 1 argument, a Connection for sending or receiving messages with the recipient.

Paxos.Transports.Common.listenOn — Method

Listen for incoming connections and invoke the indicated function when they appear. Returns a Listener

Paxos.Transports.Common.receiveFrom — Method

Receive a message over the indicated connection

Paxos.Transports.Common.sendTo — Method

Send a message on the indicated connection, using the specified Transport

Macros

Types

Constants

Index