Contents

Contents

Functions

Macros

Types

Constants

Index

Functions

Paxos.Utils.bounded — Method

Execute the provided expression fn in a separate task, waiting at most timeout seconds (using sleep). If the task does not complete in the alloted time, then abort the task and throw a TimeoutException to the caller of this function.

Paxos.Utils.compact — Method

Return an iterator that skips any nothing results from the underlying itetator

Paxos.Utils.groupBy — Function

Return a Dict whose keys are the unique values of the supplied function fn, and the values are those items in the original collection for which the supplied function returned that result.

Paxos.Utils.identity — Method

Simple identity function: returns its arguments

Paxos.Utils.maxBy — Function

Return the item with the maximum value by comparising the value returned by the supplied function fn against other values in the collection using <

Paxos.Utils.readAvailable — Method

Reads all available messages on the Channel, returning an array

Macros

Types

Paxos.Utils.TimeoutException — Type

An exception thrown if a block invoked with bounded did not complete in the specified time.

Constants

Index

Paxos.Utils.TimeoutException

Paxos.Utils.bounded

Paxos.Utils.compact

Paxos.Utils.groupBy

Paxos.Utils.identity

Paxos.Utils.maxBy

Paxos.Utils.readAvailable